py-srv/bin/requirements.txt
python-multipart 
fastapi 
uvicorn 
pydantic
tortoise-orm
asyncmy
mysql-connector-python
